so close to getting img change

+++ cbtl_base.py
@@ -65,7 +65,7 @@ class cowboytimeline:
         thumbs = []
         for v in self.tl["frames"]:
             n = v
-            n.thumbnail((400,(self.tl["mdata"]["size"][1]*((self.tl["mdata"]["size"][0]/400) ) )) )
+            n.thumbnail((400,int(self.tl["mdata"]["size"][1]*((self.tl["mdata"]["size"][0]/400) ) )) )
             thumbs.append(n)
         return thumbs
